KeysBackup: move fragments to subpackage
