+++ vector/src/main/java/im/vector/fragments/keysbackup/setup/KeysBackupSetupStep1Fragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.fragments.keysbackupsetup
+package im.vector.fragments.keysbackup.setup
 
 import butterknife.OnClick
 import im.vector.R
