+++ vector/src/main/java/im/vector/fragments/keysbackup/setup/KeysBackupSetupStep2Fragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.fragments.keysbackupsetup
+package im.vector.fragments.keysbackup.setup
 
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
