+++ vector/src/main/res/layout/fragment_keys_backup_setup_step2.xml
@@ -10,7 +10,7 @@
         android:id="@+id/keys_backup_setup"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".fragments.keysbackupsetup.KeysBackupSetupStep2Fragment">
+        tools:context=".fragments.keysbackup.setup.KeysBackupSetupStep2Fragment">
 
         <TextView
             android:id="@+id/keys_backup_setup_step2_text"
